Add tabs for sync/async example (#17)
* Add tabs for sync/async example

* Print pyright version

* Fix several build issues

Remove flake8. Use black --check instead.
Reformat files that had to.
Add Optional[] type hint for params that can be None. Pyright changed
its default behavior lately and this is a good thing.

* Fix doc build

Missed tab activation.

* Add requirements for doc building

We need sphinx-tabs to be installed.

* Fix obvious mistake in examples

* Document sync/async handlers

* Show async handler first

This is the best option, so it is better if readers see them by default.

* Add blank lines for clarity